refactor: move to internal directory
