+++ internal/timers.test.ts
