+++ test_stream.ts
@@ -14,8 +14,8 @@ import {
   OperationNotPermittedError,
   TestStreamError,
 } from "./errors/mod.ts";
-import { deferred } from "./deferred.ts";
-import { FakeTime } from "./timers.ts";
+import { deferred } from "./internal/deferred.ts";
+import { FakeTime } from "./internal/timers.ts";
 import type {
   TestStreamArgs,
   TestStreamDefinition,
@@ -27,7 +27,7 @@ import type {
   TestStreamHelperRunFn,
   TestStreamHelperWritable,
 } from "./types.ts";
-import type { MutableTuple } from "./internal_types.ts";
+import type { MutableTuple } from "./internal/types.ts";
 
 const DEFAULT_TICK_TIME = 100;
 const DEFAULT_MAX_TICKS = 50;
